Recreated entities and controllers that were affected, fixed all comments and controllers to be properly set up

+++ src/main/java/persistence/controllers/NewsfeedJpaController.java
@@ -1,17 +1,20 @@
 package persistence.controllers;
 
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import persistence.entities.FrontPageSettings;
+import java.util.ArrayList;
 import java.util.List;
 import javax.annotation.Resource;
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Named;
 import javax.persistence.EntityManager;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceContext;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 import javax.transaction.UserTransaction;
+import persistence.controllers.exceptions.IllegalOrphanException;
 import persistence.controllers.exceptions.NonexistentEntityException;
 import persistence.controllers.exceptions.RollbackFailureException;
 import persistence.entities.Newsfeed;
@@ -34,9 +37,27 @@ public class NewsfeedJpaController implements Serializable {
     }
 
     public void create(Newsfeed newsfeed) throws RollbackFailureException, Exception {
+        if (newsfeed.getFrontPageSettingsList() == null) {
+            newsfeed.setFrontPageSettingsList(new ArrayList<FrontPageSettings>());
+        }
         try {
             utx.begin();
+            List<FrontPageSettings> attachedFrontPageSettingsList = new ArrayList<FrontPageSettings>();
+            for (FrontPageSettings frontPageSettingsListFrontPageSettingsToAttach : newsfeed.getFrontPageSettingsList()) {
+                frontPageSettingsListFrontPageSettingsToAttach = em.getReference(frontPageSettingsListFrontPageSettingsToAttach.getClass(), frontPageSettingsListFrontPageSettingsToAttach.getId());
+                attachedFrontPageSettingsList.add(frontPageSettingsListFrontPageSettingsToAttach);
+            }
+            newsfeed.setFrontPageSettingsList(attachedFrontPageSettingsList);
             em.persist(newsfeed);
+            for (FrontPageSettings frontPageSettingsListFrontPageSettings : newsfeed.getFrontPageSettingsList()) {
+                Newsfeed oldNewsfeedIdOfFrontPageSettingsListFrontPageSettings = frontPageSettingsListFrontPageSettings.getNewsfeedId();
+                frontPageSettingsListFrontPageSettings.setNewsfeedId(newsfeed);
+                frontPageSettingsListFrontPageSettings = em.merge(frontPageSettingsListFrontPageSettings);
+                if (oldNewsfeedIdOfFrontPageSettingsListFrontPageSettings != null) {
+                    oldNewsfeedIdOfFrontPageSettingsListFrontPageSettings.getFrontPageSettingsList().remove(frontPageSettingsListFrontPageSettings);
+                    oldNewsfeedIdOfFrontPageSettingsListFrontPageSettings = em.merge(oldNewsfeedIdOfFrontPageSettingsListFrontPageSettings);
+                }
+            }
             utx.commit();
         } catch (Exception ex) {
             try {
@@ -48,10 +69,43 @@ public class NewsfeedJpaController implements Serializable {
         }
     }
 
-    public void edit(Newsfeed newsfeed) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void edit(Newsfeed newsfeed) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         try {
             utx.begin();
+            Newsfeed persistentNewsfeed = em.find(Newsfeed.class, newsfeed.getId());
+            List<FrontPageSettings> frontPageSettingsListOld = persistentNewsfeed.getFrontPageSettingsList();
+            List<FrontPageSettings> frontPageSettingsListNew = newsfeed.getFrontPageSettingsList();
+            List<String> illegalOrphanMessages = null;
+            for (FrontPageSettings frontPageSettingsListOldFrontPageSettings : frontPageSettingsListOld) {
+                if (!frontPageSettingsListNew.contains(frontPageSettingsListOldFrontPageSettings)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain FrontPageSettings " + frontPageSettingsListOldFrontPageSettings + " since its newsfeedId field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<FrontPageSettings> attachedFrontPageSettingsListNew = new ArrayList<FrontPageSettings>();
+            for (FrontPageSettings frontPageSettingsListNewFrontPageSettingsToAttach : frontPageSettingsListNew) {
+                frontPageSettingsListNewFrontPageSettingsToAttach = em.getReference(frontPageSettingsListNewFrontPageSettingsToAttach.getClass(), frontPageSettingsListNewFrontPageSettingsToAttach.getId());
+                attachedFrontPageSettingsListNew.add(frontPageSettingsListNewFrontPageSettingsToAttach);
+            }
+            frontPageSettingsListNew = attachedFrontPageSettingsListNew;
+            newsfeed.setFrontPageSettingsList(frontPageSettingsListNew);
             newsfeed = em.merge(newsfeed);
+            for (FrontPageSettings frontPageSettingsListNewFrontPageSettings : frontPageSettingsListNew) {
+                if (!frontPageSettingsListOld.contains(frontPageSettingsListNewFrontPageSettings)) {
+                    Newsfeed oldNewsfeedIdOfFrontPageSettingsListNewFrontPageSettings = frontPageSettingsListNewFrontPageSettings.getNewsfeedId();
+                    frontPageSettingsListNewFrontPageSettings.setNewsfeedId(newsfeed);
+                    frontPageSettingsListNewFrontPageSettings = em.merge(frontPageSettingsListNewFrontPageSettings);
+                    if (oldNewsfeedIdOfFrontPageSettingsListNewFrontPageSettings != null && !oldNewsfeedIdOfFrontPageSettingsListNewFrontPageSettings.equals(newsfeed)) {
+                        oldNewsfeedIdOfFrontPageSettingsListNewFrontPageSettings.getFrontPageSettingsList().remove(frontPageSettingsListNewFrontPageSettings);
+                        oldNewsfeedIdOfFrontPageSettingsListNewFrontPageSettings = em.merge(oldNewsfeedIdOfFrontPageSettingsListNewFrontPageSettings);
+                    }
+                }
+            }
             utx.commit();
         } catch (Exception ex) {
             try {
@@ -70,7 +124,7 @@ public class NewsfeedJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         try {
             utx.begin();
             Newsfeed newsfeed;
@@ -79,6 +133,17 @@ public class NewsfeedJpaController implements Serializable {
                 newsfeed.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The newsfeed with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<FrontPageSettings> frontPageSettingsListOrphanCheck = newsfeed.getFrontPageSettingsList();
+            for (FrontPageSettings frontPageSettingsListOrphanCheckFrontPageSettings : frontPageSettingsListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Newsfeed (" + newsfeed + ") cannot be destroyed since the FrontPageSettings " + frontPageSettingsListOrphanCheckFrontPageSettings + " in its frontPageSettingsList field has a non-nullable newsfeedId field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(newsfeed);
             utx.commit();
